pyproject.toml
[project]
name = "pixeletica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litemapy>=0.10.0b0",
    "pillow>=11.1.0",
]
